Make all the tests pass!!!
But in a terrible way.

We've reached the end of the usefulness of the current test suite.

Which is actually awesome, because it means our parser isn't raising
errors willy-nilly. It has a too-greedy matcher in the action rule. But
that's OK because that's sort of the design of the Fountain spec, I
think: anything that isn't identifiable as some specific element is
probably action.

The next step is to rewrite or extend the tests to assert that dialogue
is dialogue, not action.