Introduce emphasized text
Misc:

* allow rspec failure re-running
* add some classes for italics and plain text phrases
* introduce some equality comparisons for screenplay elements

To that last point, the idea there is that the objects are simple value
objects, and if there are two simple value objects with the same value,
they should be considered equal. This makes it easier to test them,
primarily.

Emphasized text is mostly working, but it's not perfect. I think the
approach may be correct. It feels weird. I don't know.